pyproject.toml
[tool.poetry]
authors = ["Tim <tim.shawn@gmail.com>"]
description = "This example starts up a FastAPI server"
license = "MIT"
name = "app"
version = "0.1.0"

[tool.poetry.dependencies]
faiss-cpu = "^1.7.3"
fastapi = "^0.80.0"
langchain = "0.0.77"
openai = "^0.26.4"
python = "^3.11"
uvicorn = "^0.18.3"
python-dotenv = "^0.21.1"

[tool.poetry.scripts]
start = "app.main:start"

[tool.poetry.dev-dependencies]

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
